Inserts somewhat working.

+++ Source/Profiling/SisoDb.Profiling/Model/Customer.cs
@@ -5,7 +5,7 @@ namespace SisoDb.Profiling.Model
     [Serializable]
     public class Customer
     {
-        public int StructureId { get; set; }
+        public long StructureId { get; set; }
 
         public int CustomerNo { get; set; }
